Add Bool32 type to work around go -race false positives
go -race is producing false positives because it sees
Load/StoreBool() accessing a uint32, even though only one of the
bytes is used. When other variables which happen to be in the other 3 bytes
are accessed then go -race complains that our uint32 access races.

By keeping the other 3 bytes free of other users we avoid this. Hence
 type Bool32 uint32